API

Public

WebDriver.WebDriver — Module

WebDriver

A binding for the WebDriver and JSON Wire Protocol for the Julia programing language compatible with the Selenium WebDriver.

Repository

Example

julia> using WebDriver

WebDriver.Capabilities — Type

Capabilities(browserName::AbstractString;
             browserVersion::AbstractString = "",
             platformName::AbstractString = "",
             acceptInsecureCerts::Union{Nothing, Bool} = nothing,
             pageLoadStrategy::AbstractString = "",
             proxy::Union{Nothing, Proxy} = nothing,
             setWindowRect::Union{Nothing, Bool} = nothing,
             timeouts::Timeouts = Timeouts(),
             strictFileInteractability::Union{Nothing, Bool} = nothing,
             unhandledPromptBehavior::AbstractString = "dismiss and notify"
             )::Capabilities

According to W3C.

Example

julia> capabilities = Capabilities("chrome")
Remote WebDriver Capabilities
  browserName: chrome
  Session Timeouts -- script: 30000, pageLoad: 300000, implicit: 0
  unhandledPromptBehavior: dismiss and notify

WebDriver.Cookie — Type

Cookie(obj)::Cookie
Cookie(name::AbstractString, value;
       path::AbstractString = "/",
       domain::Union{Nothing, AbstractString} = nothing,
       secure::Bool = false,
       httpOnly::Bool = false,
       expiry::Union{Nothing, Integer, DateTime} = nothing)::Cookie

A cookie per W3C.

WebDriver.Element — Type

Element(session::Session, location_strategy::AbstractString, value::AbstractString)::Element
Element(element::Element, location_strategy::AbstractString, value::AbstractString)::Element

Finds an element based on a session or element. May use a location strategy from "css selector", "link text", "partial link text", "tag name" or "xpath".

WebDriver.Proxy — Type

Proxy(proxyType::AbstractString;
      proxyAutoconfigUrl::Union{Nothing, AbstractString} = nothing,
      ftpProxy::Union{Nothing, AbstractString} = nothing,
      httpProxy::Union{Nothing, AbstractString} = nothing,
      noProxy::Union{Nothing, AbstractVector{<:AbstractString}} = nothing,
      sslProxy::Union{Nothing, AbstractString} = nothing,
      socksProxy::Union{Nothing, AbstractString} = nothing,
      socksVersion::Union{Nothing, Integer} = nothing)::Proxy

The proxy configuration capability is a JSON Object nested within the primary capabilities. According to the W3C.

WebDriver.RemoteWebDriver — Type

RemoteWebDriver(capabilities::Capabilities;
                host::AbstractString = "localhost",
                port::Integer = 4444,
                path::AbstractString = "/wd/hub",
				kwargs...)::RemoteWebDriver

Specifies a remote web driver according to W3C.

julia> capabilities = Capabilities("chrome")
Remote WebDriver Capabilities
  browserName: chrome
  Session Timeouts -- script: 30000, pageLoad: 300000, implicit: 0
  unhandledPromptBehavior: dismiss and notify
julia> wd = RemoteWebDriver(capabilities, host = "selenium")
RemoteWebDriver{Capabilities{Nothing,Nothing,Nothing,Nothing}}("http://selenium:4444/wd/hub", Remote WebDriver Capabilities
  browserName: chrome
  Session Timeouts -- script: 30000, pageLoad: 300000, implicit: 0
  unhandledPromptBehavior: dismiss and notify
, Dict{String,Union{Int64, String}}())

WebDriver.Session — Type

Session

This is a web session.

julia> capabilities = Capabilities("chrome")
Remote WebDriver Capabilities
  browserName: chrome
  Session Timeouts -- script: 30000, pageLoad: 300000, implicit: 0
  unhandledPromptBehavior: dismiss and notify
julia> wd = RemoteWebDriver(capabilities, port = parse(Int, ENV["WEBDRIVER_PORT"]))
RemoteWebDriver{Capabilities{Nothing,Nothing,Nothing,Nothing}}("http://localhost:32768/wd/hub", Remote WebDriver Capabilities
  browserName: chrome
  Session Timeouts -- script: 30000, pageLoad: 300000, implicit: 0
  unhandledPromptBehavior: dismiss and notify
, Dict{String,Union{Int64, String}}())
julia> session = Session(wd);

julia> isa(session, Session)
true

WebDriver.Timeouts — Type

Timeouts

According to the W3C.

julia> Timeouts(script = 50_000, pageLoad = 100_000, implicit = 5)
Session Timeouts -- script: 50000, pageLoad: 100000, implicit: 5

WebDriver.maximize! — Method

maximize!(session::Session; window::AbstractString = "current")::Dict{String,Int}

Maximize the window based on window handle (defaults to current). Returns Rect oof window.

Note

This command is implemented using the JSON Wire Protocol.

WebDriver.moveto! — Method

moveto!(session::Session; x::Integer = 0, y::Integer = 0)
moveto!(element::Element; x::Integer = 0, y::Integer = 0)

WebDriver.rect — Function

rect(session::Session; window::AbstractString = "current")::NamedTuple

Return window's dimensions in pixels.

WebDriver.rect! — Method

rect!(session::Session;
      window::AbstractString = "current",
      width::Union{Nothing, Real} = nothing,
      height::Union{Nothing, Real} = nothing
     )::NamedTuple

Sets the window's dimensions in pixels.

WebDriver.rect — Method

rect(element::Element)::NamedTuple

Return the size of the element in pixels.

WebDriver.script — Method

script(session::Session, script::AbstractString, args...; async::Bool = false)

Executes JavaScript (sync / async)

Note

This command uses the JSON Wire Protocol instead of the current W3c WebDriver API.

WebDriver.status — Method

status(wd::WebDriver)::Bool

Returns true if the server is running and ready to accept new sessions.

WebDriver.timeouts — Method

timeouts(session::Session)

Get the timeouts according to webdriver2. Currently not supported by Selenium.

Private

Base.delete! — Method

delete!(session::Session, cookie::AbstractString)

If cookie = "", delete all cookies. Otherwise, delete cookie matching the "cookie" name.

Base.delete! — Method

delete!(session::Session)

Deletes the session from the Remote Driver.

WebDriver.actions! — Method

actions!(session::Session, actions::AbstractVector)